:hammer: refactor parsing on property
This commit is to refactor method of ParseOnEvent which check strict types.
